Set CI=true for npm and gatsby build
It seems that this affects gatsby's output. I believe it also has an
effect on npm so perhaps it's generally good practice when using it with
both these tools.

I also briefly saw it mentioned that it might help with some performance
so perhaps it'll also help/fix the issues with the build freezing up.