Disabling npm audit for now
I believe because of the fact that gatsby statically generates sites,
most audits aren't really relevant. Although I would still like to fix
them, it seems that a lot of packages in the gatsby ecosystem don't
prioritize this? In any case, I don't think there's too much value in
keeping this as part of the build.